+++ lib/massive_record/orm/relations/interface.rb
@@ -102,7 +102,7 @@ module MassiveRecord
           def create_references_one_accessors(metadata)
             redefine_method(metadata.name) do
               proxy = relation_proxy(metadata.name)
-              proxy.load_target ? proxy : nil
+              proxy.load_proxy_target ? proxy : nil
             end
 
             redefine_method(metadata.name+'=') do |record|