+++ lib/massive_record/orm/relations/proxy.rb
@@ -4,7 +4,7 @@ module MassiveRecord
       #
       # Parent class for all proxies sitting between records.
       # It's responsibility is to transparently load and forward
-      # method calls to it's target. Iy may also do some small maintenance
+      # method calls to it's proxy_target. Iy may also do some small maintenance
       # work like setting foreign key in proxy_owner object etc. That kind of
       # functionality is likely to be implemented in one of it's more
       # specific sub class proxies.
@@ -12,76 +12,76 @@ module MassiveRecord
       class Proxy
         instance_methods.each { |m| undef_method m unless m.to_s =~ /^(?:nil\?|send|object_id|to_a|to_s|extend|equal\?)$|^__|^respond_to|^should|^instance_variable_/ }
 
-        attr_reader :target
+        attr_reader :proxy_target
         attr_accessor :proxy_owner, :metadata
 
         delegate :foreign_key, :foreign_key_setter, :store_in, :store_foreign_key_in,
           :polymorphic_type_column, :polymorphic_type_column_setter,
-          :class_name, :target_class, :name, :persisting_foreign_key?, :find_with,
+          :class_name, :proxy_target_class, :name, :persisting_foreign_key?, :find_with,
           :represents_a_collection?, :to => :metadata
 
         def initialize(options = {})
           options.to_options!
           self.metadata = options[:metadata]
           self.proxy_owner = options[:proxy_owner]
-          self.target = options[:target]
+          self.proxy_target = options[:proxy_target]
 
-          reset if target.nil?
+          reset if proxy_target.nil?
         end
 
 
         #
-        # The target of a relation is the record
+        # The proxy_target of a relation is the record
         # the proxy_owner references. For instance,
         # 
         # class Person
         #   references_one :car
         # end
         #
-        # The proxy_owner is a record of class person, the target will be the car.
+        # The proxy_owner is a record of class person, the proxy_target will be the car.
         #
-        def target=(target)
-          @target = target
-          loaded! unless @target.nil?
+        def proxy_target=(proxy_target)
+          @proxy_target = proxy_target
+          loaded! unless @proxy_target.nil?
         end
         
         #
-        # Returns the target. Loads it, if it's not there.
-        # Returns nil if for some reason target could not be found.
+        # Returns the proxy_target. Loads it, if it's not there.
+        # Returns nil if for some reason proxy_target could not be found.
         #
-        def load_target
-          self.target = find_target_or_find_with_proc if find_target?
-          target
+        def load_proxy_target
+          self.proxy_target = find_proxy_target_or_find_with_proc if find_proxy_target?
+          proxy_target
         rescue RecordNotFound
           reset
         end
 
         def reload
           reset
-          load_target
+          load_proxy_target
         end
 
         def reset
-          @loaded = @target = nil
+          @loaded = @proxy_target = nil
         end
 
-        def replace(target)
-          if target.nil?
+        def replace(proxy_target)
+          if proxy_target.nil?
             reset 
           else
-            raise_if_type_mismatch(target)
-            self.target = target
+            raise_if_type_mismatch(proxy_target)
+            self.proxy_target = proxy_target
           end
         end
 
         def inspect
-          load_target.inspect
+          load_proxy_target.inspect
         end
 
 
 
         #
-        # If the proxy is loaded it has a target
+        # If the proxy is loaded it has a proxy_target
         #
         def loaded?
           !!@loaded
@@ -95,14 +95,14 @@ module MassiveRecord
 
 
         def respond_to?(*args)
-          super || (load_target && target.respond_to?(*args))
+          super || (load_proxy_target && proxy_target.respond_to?(*args))
         end
 
         def method_missing(method, *args, &block)
-          return target.send(method, *args, &block) if load_target && target.respond_to?(method)
+          return proxy_target.send(method, *args, &block) if load_proxy_target && proxy_target.respond_to?(method)
           super
         rescue NoMethodError => e
-          raise e, e.message.sub(/ for #<.*$/, " via proxy for #{target}")
+          raise e, e.message.sub(/ for #<.*$/, " via proxy for #{proxy_target}")
         end
       
 
@@ -110,60 +110,60 @@ module MassiveRecord
         #           With Rails it seems like the proxy answered to to_param, which
         #           kinda was not what I wanted.
         def to_param # :nodoc:
-          target.try :to_param
+          proxy_target.try :to_param
         end
         
 
         protected
 
-        def find_target_or_find_with_proc
-          find_with_proc? ? find_target_with_proc : find_target
+        def find_proxy_target_or_find_with_proc
+          find_with_proc? ? find_proxy_target_with_proc : find_proxy_target
         end
 
         #
-        # "Abstract" method used to find target for the proxy.
+        # "Abstract" method used to find proxy_target for the proxy.
         # Implement in subclasses. It is not called when the meta
-        # data contains a find_with proc; in that case find_target_with_proc
+        # data contains a find_with proc; in that case find_proxy_target_with_proc
         # is used instead
         #
-        def find_target
+        def find_proxy_target
         end
         
         #
         # Gives sub classes a place to hook into when we are
-        # gonna find target(s) by the proc. For instance, the
+        # gonna find proxy_target(s) by the proc. For instance, the
         # references_many proxy ensures that the result of proc
         # is put inside of an array.
         #
-        def find_target_with_proc(options = {})
+        def find_proxy_target_with_proc(options = {})
           find_with.call(proxy_owner, options)
         end
 
         #
-        # When are we supposed to find a target? Find a target is done
-        # through load_target.
+        # When are we supposed to find a proxy_target? Find a proxy_target is done
+        # through load_proxy_target.
         #
-        def find_target?
-          !loaded? && can_find_target?
+        def find_proxy_target?
+          !loaded? && can_find_proxy_target?
         end
 
         #
-        # Override this to controll when a target may be found.
+        # Override this to controll when a proxy_target may be found.
         #
-        def can_find_target?
+        def can_find_proxy_target?
           find_with_proc?
         end
 
         #
-        # Are we supposed to find target with a proc?
+        # Are we supposed to find proxy_target with a proc?
         #
         def find_with_proc?
           !find_with.nil? && find_with.respond_to?(:call)
         end
 
         def raise_if_type_mismatch(record)
-          unless record.is_a? target_class
-            message = "#{class_name}(##{target_class.object_id}) expected, got #{record.class}(##{record.class.object_id})"
+          unless record.is_a? proxy_target_class
+            message = "#{class_name}(##{proxy_target_class.object_id}) expected, got #{record.class}(##{record.class.object_id})"
             raise RelationTypeMismatch.new(message)
           end
         end