+++ spec/orm/relations/interface_spec.rb
@@ -55,35 +55,35 @@ describe MassiveRecord::ORM::Relations::Interface do
           subject.boss.should be_nil 
         end
 
-        it "should return the proxy's target if boss is set" do
+        it "should return the proxy's proxy_target if boss is set" do
           subject.boss = boss
           subject.boss.should == boss
         end
 
         it "should be able to reset the proxy" do
-          proxy.should_receive(:load_target).and_return(true)
+          proxy.should_receive(:load_proxy_target).and_return(true)
           proxy.should_receive(:reset) 
           subject.boss.reset
         end
 
         it "should be able to reload the proxy" do
-          proxy.should_receive(:load_target).and_return(true)
+          proxy.should_receive(:load_proxy_target).and_return(true)
           proxy.should_receive(:reload)
           subject.boss.reload
         end
 
-        it "should set the foreign_key in proxy_owner when target is set" do
+        it "should set the foreign_key in proxy_owner when proxy_target is set" do
           subject.boss = boss
           subject.boss_id.should == boss.id
         end
 
-        it "should load target object when read method is called" do
+        it "should load proxy_target object when read method is called" do
           PersonWithTimestamp.should_receive(:find).and_return(boss)
           subject.boss_id = boss.id
           subject.boss.should == boss
         end
 
-        it "should not load target twice" do
+        it "should not load proxy_target twice" do
           PersonWithTimestamp.should_receive(:find).once.and_return(boss)
           subject.boss_id = boss.id
           2.times { subject.boss }
@@ -128,17 +128,17 @@ describe MassiveRecord::ORM::Relations::Interface do
           subject.attachable.should be_nil 
         end
 
-        it "should return the proxy's target if attachable is set" do
+        it "should return the proxy's proxy_target if attachable is set" do
           subject.attachable = attachable
           subject.attachable.should == attachable
         end
 
-        it "should set the foreign_key in proxy_owner when target is set" do
+        it "should set the foreign_key in proxy_owner when proxy_target is set" do
           subject.attachable = attachable
           subject.attachable_id.should == attachable.id
         end
 
-        it "should set the type in proxy_owner when target is set" do
+        it "should set the type in proxy_owner when proxy_target is set" do
           subject.attachable = attachable
           subject.attachable_type.should == attachable.class.to_s.underscore
         end
@@ -154,12 +154,12 @@ describe MassiveRecord::ORM::Relations::Interface do
               subject.attachable_type = polymorphic_class.to_s.underscore
             end
 
-            it "should load target object when read method is called" do
+            it "should load proxy_target object when read method is called" do
               polymorphic_class.should_receive(:find).and_return(attachable)
               subject.attachable.should == attachable
             end
 
-            it "should not load target twice" do
+            it "should not load proxy_target twice" do
               polymorphic_class.should_receive(:find).once.and_return(attachable)
               2.times { subject.attachable }
             end