+++ spec/shared/orm/relations/proxy.rb
@@ -1,21 +1,21 @@
 shared_examples_for "relation proxy" do
-  let(:target) { mock(Object).as_null_object }
-  let(:find_target_returns) { subject.represents_a_collection? ? [target] : target }
+  let(:proxy_target) { mock(Object).as_null_object }
+  let(:find_proxy_target_returns) { subject.represents_a_collection? ? [proxy_target] : proxy_target }
 
   before do
     subject.metadata = mock(MassiveRecord::ORM::Relations::Metadata, :find_with => nil).as_null_object if subject.metadata.nil?
   end
 
-  %w(proxy_owner target metadata).each do |method|
+  %w(proxy_owner proxy_target metadata).each do |method|
     it "should respond to #{method}" do
       should respond_to method
     end
   end
 
   it "should be setting values by initializer" do
-    proxy = MassiveRecord::ORM::Relations::Proxy.new(:proxy_owner => "proxy_owner", :target => target, :metadata => "metadata")
+    proxy = MassiveRecord::ORM::Relations::Proxy.new(:proxy_owner => "proxy_owner", :proxy_target => proxy_target, :metadata => "metadata")
     proxy.proxy_owner.should == "proxy_owner"
-    proxy.target.should == target
+    proxy.proxy_target.should == proxy_target
     proxy.metadata.should == "metadata"
   end
 
@@ -45,17 +45,17 @@ shared_examples_for "relation proxy" do
       should_not be_loaded
     end
 
-    it "should reset the target" do
-      subject.target = target
+    it "should reset the proxy_target" do
+      subject.proxy_target = proxy_target
       subject.reset
-      subject.stub(:find_target?).and_return(false)
-      subject.target.should be_blank
+      subject.stub(:find_proxy_target?).and_return(false)
+      subject.proxy_target.should be_blank
     end
   end
 
   describe "#reload" do
     before do
-      subject.stub!(:find_target)
+      subject.stub!(:find_proxy_target)
     end
 
     it "should reset the proxy" do
@@ -63,20 +63,20 @@ shared_examples_for "relation proxy" do
       subject.reload
     end
 
-    it "should call load_target" do
-      subject.should_receive :load_target
+    it "should call load_proxy_target" do
+      subject.should_receive :load_proxy_target
       subject.reload
     end
 
-    it "should return target if loaded successfully" do
-      subject.should_receive(:can_find_target?).and_return true
-      subject.should_receive(:find_target) { find_target_returns }
-      subject.reload.should == find_target_returns
+    it "should return proxy_target if loaded successfully" do
+      subject.should_receive(:can_find_proxy_target?).and_return true
+      subject.should_receive(:find_proxy_target) { find_proxy_target_returns }
+      subject.reload.should == find_proxy_target_returns
     end
 
-    it "should return nil if loading of target failed" do
-      subject.stub(:can_find_target?).and_return true
-      subject.should_receive(:find_target).and_raise MassiveRecord::ORM::RecordNotFound
+    it "should return nil if loading of proxy_target failed" do
+      subject.stub(:can_find_proxy_target?).and_return true
+      subject.should_receive(:find_proxy_target).and_raise MassiveRecord::ORM::RecordNotFound
       subject.reload.should be_blank
     end
   end
@@ -84,49 +84,49 @@ shared_examples_for "relation proxy" do
 
 
 
-  describe "target" do
-    it "should return the target if it is present" do
-      subject.target = target
-      subject.target.should == target
+  describe "proxy_target" do
+    it "should return the proxy_target if it is present" do
+      subject.proxy_target = proxy_target
+      subject.proxy_target.should == proxy_target
     end
 
-    it "should be consodered loaded if a target has been set" do
-      subject.target = target
+    it "should be consodered loaded if a proxy_target has been set" do
+      subject.proxy_target = proxy_target
       should be_loaded
     end
 
-    it "should not try to load target if it has been loaded" do
+    it "should not try to load proxy_target if it has been loaded" do
       subject.loaded!
-      should_not_receive :find_target
-      subject.load_target.should be_blank
+      should_not_receive :find_proxy_target
+      subject.load_proxy_target.should be_blank
     end
 
-    it "should try to load the target if it has not been loaded" do
-      subject.stub(:can_find_target?).and_return true
-      subject.should_receive(:find_target) { find_target_returns }
-      subject.load_target
-      subject.target.should == find_target_returns
+    it "should try to load the proxy_target if it has not been loaded" do
+      subject.stub(:can_find_proxy_target?).and_return true
+      subject.should_receive(:find_proxy_target) { find_proxy_target_returns }
+      subject.load_proxy_target
+      subject.proxy_target.should == find_proxy_target_returns
     end
 
-    it "should reset proxy if target's record was not found" do
-      subject.stub(:can_find_target?).and_return true
-      subject.should_receive(:find_target).and_raise MassiveRecord::ORM::RecordNotFound
-      subject.load_target.should be_blank
+    it "should reset proxy if proxy_target's record was not found" do
+      subject.stub(:can_find_proxy_target?).and_return true
+      subject.should_receive(:find_proxy_target).and_raise MassiveRecord::ORM::RecordNotFound
+      subject.load_proxy_target.should be_blank
     end
   end
 
 
   describe "replace" do
-    let(:old_target) { subject.represents_a_collection? ? [subject.target_class.new] : subject.target_class.new }
-    let(:new_target) { subject.represents_a_collection? ? [subject.target_class.new] : subject.target_class.new }
+    let(:old_proxy_target) { subject.represents_a_collection? ? [subject.proxy_target_class.new] : subject.proxy_target_class.new }
+    let(:new_proxy_target) { subject.represents_a_collection? ? [subject.proxy_target_class.new] : subject.proxy_target_class.new }
 
     before do
-      subject.target = old_target
+      subject.proxy_target = old_proxy_target
     end
 
-    it "should replace the old target with the new one" do
-      subject.replace(new_target)
-      subject.target.should == new_target
+    it "should replace the old proxy_target with the new one" do
+      subject.replace(new_proxy_target)
+      subject.proxy_target.should == new_proxy_target
     end
 
     it "should reset the proxy if asked to replace with nil" do
@@ -136,19 +136,19 @@ shared_examples_for "relation proxy" do
   end
 
   describe "find_with" do
-    let(:metadata) { MassiveRecord::ORM::Relations::Metadata.new 'person', :find_with => Proc.new { |target| Person.find("testing-123") }}
+    let(:metadata) { MassiveRecord::ORM::Relations::Metadata.new 'person', :find_with => Proc.new { |proxy_target| Person.find("testing-123") }}
     let(:person) { Person.new }
 
     before do
       subject.metadata = metadata
-      subject.stub(:can_find_target?).and_return(true)
+      subject.stub(:can_find_proxy_target?).and_return(true)
     end
 
     it "should use metadata's find with if exists" do
       Person.should_receive(:find).with("testing-123").and_return(person)
 
-      should_not_receive :find_target
-      subject.load_target 
+      should_not_receive :find_proxy_target
+      subject.load_proxy_target 
     end
   end
 end